feat: add message constants folder on fds_shared
