+++ src/fsd_shared/configs/constants/index.ts
@@ -1,0 +1,1 @@
+export { MESSAGES } from './messages';
