+++ src/fsd_shared/configs/constants/messages.ts
@@ -1,0 +1,3 @@
+export const MESSAGES = Object.freeze({
+  PREVIOUS_BUTTON_TEXT: '이전',
+});
